Refactor: Decompose monolithic page components and extract hooks
