+++ client/src/components/meetings/PeerVideo.jsx
@@ -1,0 +1,28 @@
+import React, { useEffect, useRef } from "react";
+
+export default function PeerVideo({ peer, userInfo }) {
+  const ref = useRef();
+
+  useEffect(() => {
+    peer.on("stream", (stream) => {
+      if (ref.current) {
+        ref.current.srcObject = stream;
+      }
+    });
+  }, [peer]);
+
+  return (
+    <div className="relative bg-black rounded-2xl overflow-hidden shadow-lg aspect-video flex-1 min-w-[280px] max-w-[600px] border border-gray-800">
+      <video
+        playsInline
+        autoPlay
+        ref={ref}
+        className="w-full h-full object-cover"
+      />
+      <div className="absolute bottom-4 left-4 bg-black/60 px-3 py-1.5 rounded-lg backdrop-blur-sm text-white text-sm flex items-center gap-2">
+        <div className="w-2 h-2 rounded-full bg-green-500 animate-pulse" />
+        {userInfo?.name || "Participant"}
+      </div>
+    </div>
+  );
+}
