+++ client/src/pages/MeetingRoom.jsx
@@ -1,105 +1,61 @@
-import React, { useEffect, useRef, useState } from "react";
-import { useParams, useNavigate } from "react-router-dom";
-import { io } from "socket.io-client";
-import Peer from "simple-peer";
-import {
-  Mic,
-  MicOff,
-  PhoneOff,
-  Loader2,
-  CheckCircle2,
-  Sparkles,
-  Video,
-  VideoOff,
-  MonitorUp,
-  Users,
-  Clock,
-  Copy,
-  NotebookPen,
-  PanelRightClose,
-  Captions,
-  FileText,
-  Download,
-  Search,
-  X,
-} from "lucide-react";
+import React, { useEffect, useState } from "react";
+import { useParams } from "react-router-dom";
 import { toast } from "react-toastify";
+import { Loader2, CheckCircle2 } from "lucide-react";
 import ErrorState from "../components/ErrorState.jsx";
 import CollaborativeEditor from "../components/meetings/CollaborativeEditor.jsx";
-
-// A separate component to render each peer's video
-const PeerVideo = ({ peer, userInfo }) => {
-  const ref = useRef();
-
-  useEffect(() => {
-    peer.on("stream", (stream) => {
-      if (ref.current) {
-        ref.current.srcObject = stream;
-      }
-    });
-  }, [peer]);
-
-  return (
-    <div className="relative bg-black rounded-2xl overflow-hidden shadow-lg aspect-video flex-1 min-w-[280px] max-w-[600px] border border-gray-800">
-      <video
-        playsInline
-        autoPlay
-        ref={ref}
-        className="w-full h-full object-cover"
-      />
-      <div className="absolute bottom-4 left-4 bg-black/60 px-3 py-1.5 rounded-lg backdrop-blur-sm text-white text-sm flex items-center gap-2">
-        <div className="w-2 h-2 rounded-full bg-green-500 animate-pulse" />
-        {userInfo?.name || "Participant"}
-      </div>
-    </div>
-  );
-};
+import PeerVideo from "../components/meetings/PeerVideo.jsx";
+import MeetingHeader from "../components/meetings/MeetingHeader.jsx";
+import MeetingControlBar from "../components/meetings/MeetingControlBar.jsx";
+import TranscriptPanel from "../components/meetings/TranscriptPanel.jsx";
+import LiveCaptions from "../components/meetings/LiveCaptions.jsx";
+import useWebRTC from "../hooks/useWebRTC";
+import useLiveTranscription from "../hooks/useLiveTranscription";
 
 const MeetingRoom = () => {
   const { roomId } = useParams();
-  const navigate = useNavigate();
-
-  const [joined, setJoined] = useState(false);
-  const [loading, setLoading] = useState(false);
-  const [meetingEnded, setMeetingEnded] = useState(false);
-  const [mediaError, setMediaError] = useState(null);
-
-  const [micOn, setMicOn] = useState(true);
-  const [cameraOn, setCameraOn] = useState(true);
-  const [isScreenSharing, setIsScreenSharing] = useState(false);
-
-  const [peers, setPeers] = useState([]);
   const [duration, setDuration] = useState(0);
   const [showNotes, setShowNotes] = useState(false);
+
+  // Transcription state
   const [showCaptions] = useState(true);
   const [showTranscript, setShowTranscript] = useState(false);
   const [captions, setCaptions] = useState([]);
   const [transcriptSegments, setTranscriptSegments] = useState([]);
   const [transcriptionEnabled, setTranscriptionEnabled] = useState(false);
-  const [audioContext, setAudioContext] = useState(null);
-  const [audioProcessor, setAudioProcessor] = useState(null);
 
-  const socketRef = useRef();
-  const userVideoRef = useRef();
-  const streamRef = useRef();
-  const screenTrackRef = useRef();
-  const peersRef = useRef([]);
+  // WebRTC
+  const {
+    joined,
+    loading,
+    meetingEnded,
+    mediaError,
+    setMediaError,
+    micOn,
+    cameraOn,
+    isScreenSharing,
+    peers,
+    socketRef,
+    userVideoRef,
+    streamRef,
+    joinMeeting,
+    leaveMeeting,
+    toggleMic,
+    toggleCamera,
+    toggleScreenShare,
+  } = useWebRTC(roomId, {
+    showCaptions,
+    setCaptions,
+    setTranscriptSegments,
+    setTranscriptionEnabled,
+  });
 
-  const backendUrl = import.meta.env.VITE_BACKEND_URL;
-
-  // Format time for the duration timer
-  const formatTime = (seconds) => {
-    const hrs = Math.floor(seconds / 3600);
-    const mins = Math.floor((seconds % 3600) / 60);
-    const secs = seconds % 60;
-    return `${hrs > 0 ? hrs + ":" : ""}${mins.toString().padStart(2, "0")}:${secs.toString().padStart(2, "0")}`;
-  };
-
-  const formatTimestamp = (seconds) => {
-    const mins = Math.floor(seconds / 60);
-    const secs = Math.floor(seconds % 60);
-    return `${mins.toString().padStart(2, "0")}:${secs.toString().padStart(2, "0")}`;
-  };
+  // Transcription
+  const { toggleTranscription } = useLiveTranscription(
+    roomId,
+    socketRef,
+    streamRef,
+  );
 
   useEffect(() => {
     let timer;
@@ -111,321 +67,9 @@ const MeetingRoom = () => {
     return () => clearInterval(timer);
   }, [joined, meetingEnded]);
 
-  const joinMeeting = async () => {
-    try {
-      setLoading(true);
-
-      const stream = await navigator.mediaDevices.getUserMedia({
-        video: true,
-        audio: true,
-      });
-
-      streamRef.current = stream;
-      setJoined(true);
-
-      setTimeout(() => {
-        if (userVideoRef.current) {
-          userVideoRef.current.srcObject = stream;
-        }
-      }, 100);
-
-      socketRef.current = io(backendUrl, { transports: ["websocket"] });
-
-      const userInfo = { name: "You" }; // In a real app, grab from context
-
-      socketRef.current.emit("join-meeting", { roomId, userInfo });
-
-      socketRef.current.on("all-users", (users) => {
-        const peersArr = [];
-        users.forEach((user) => {
-          const peer = createPeer(user.socketId, socketRef.current.id, stream);
-          peersRef.current.push({
-            peerID: user.socketId,
-            peer,
-            userInfo: user,
-          });
-          peersArr.push({
-            peerID: user.socketId,
-            peer,
-            userInfo: user,
-          });
-        });
-        setPeers(peersArr);
-      });
-
-      socketRef.current.on("user-joined", (user) => {
-        toast.info(`👋 Participant joined`);
-        const peer = addPeer(user.socketId, socketRef.current.id, stream);
-        peersRef.current.push({
-          peerID: user.socketId,
-          peer,
-          userInfo: user,
-        });
-
-        setPeers([...peersRef.current]);
-      });
-
-      socketRef.current.on("user-joined-signal", (payload) => {
-        const item = peersRef.current.find(
-          (p) => p.peerID === payload.callerID,
-        );
-        if (item) {
-          item.peer.signal(payload.signal);
-        }
-      });
-
-      socketRef.current.on("receiving-returned-signal", (payload) => {
-        const item = peersRef.current.find((p) => p.peerID === payload.id);
-        if (item) {
-          item.peer.signal(payload.signal);
-        }
-      });
-
-      socketRef.current.on("user-left", (id) => {
-        toast.error(`🚪 Participant left`);
-        const peerObj = peersRef.current.find((p) => p.peerID === id);
-        if (peerObj) {
-          peerObj.peer.destroy();
-        }
-        peersRef.current = peersRef.current.filter((p) => p.peerID !== id);
-        setPeers([...peersRef.current]);
-      });
-
-      // Transcription events
-      socketRef.current.on("transcript-partial", (data) => {
-        if (showCaptions) {
-          setCaptions((prev) => [
-            ...prev.slice(-4),
-            { text: data.text, isFinal: false, timestamp: data.timestamp },
-          ]);
-        }
-      });
-
-      socketRef.current.on("transcript-final", (data) => {
-        const { segment } = data;
-        setCaptions((prev) => [
-          ...prev.slice(-4),
-          { text: segment.text, speaker: segment.speaker, isFinal: true, timestamp: data.timestamp },
-        ]);
-        setTranscriptSegments((prev) => [...prev, segment]);
-      });
-
-      socketRef.current.on("transcription-started", () => {
-        setTranscriptionEnabled(true);
-        toast.success("🎙️ Live transcription started");
-      });
-
-      socketRef.current.on("transcription-stopped", () => {
-        setTranscriptionEnabled(false);
-        toast.info("🎙️ Live transcription stopped");
-      });
-
-      socketRef.current.on("transcription-error", (data) => {
-        toast.error(`Transcription error: ${data.message}`);
-        setTranscriptionEnabled(false);
-      });
-
-      setLoading(false);
-    } catch (err) {
-      console.error("Camera/Mic access denied:", err);
-      let errMsg =
-        "Camera or microphone access denied. Please enable them and retry.";
-      if (err.name === "NotFoundError" || err.name === "DevicesNotFoundError") {
-        errMsg = "Required media devices (camera or microphone) not found.";
-      } else if (
-        err.name === "NotAllowedError" ||
-        err.name === "PermissionDeniedError"
-      ) {
-        errMsg =
-          "Permission denied. Please allow camera and microphone access in your browser settings.";
-      }
-      setMediaError(errMsg);
-      toast.error(errMsg);
-      setLoading(false);
-    }
-  };
-
-  const createPeer = (userToSignal, callerID, stream) => {
-    const peer = new Peer({
-      initiator: true,
-      trickle: false,
-      stream,
-    });
-
-    peer.on("signal", (signal) => {
-      socketRef.current.emit("sending-signal", {
-        userToSignal,
-        callerID,
-        signal,
-        userInfo: { name: "You" },
-      });
-    });
-
-    return peer;
-  };
-
-  const addPeer = (incomingSignal, callerID, stream) => {
-    const peer = new Peer({
-      initiator: false,
-      trickle: false,
-      stream,
-    });
-
-    peer.on("signal", (signal) => {
-      socketRef.current.emit("returning-signal", { signal, callerID });
-    });
-
-    return peer;
-  };
-
-  const leaveMeeting = () => {
-    setMeetingEnded(true);
-
-    streamRef.current?.getTracks().forEach((track) => track.stop());
-    screenTrackRef.current?.getTracks().forEach((track) => track.stop());
-
-    socketRef.current?.disconnect();
-
-    setJoined(false);
-
-    setTimeout(() => {
-      setMeetingEnded(false);
-      navigate("/dashboard");
-    }, 4000);
-  };
-
-  // Toggle Media Handlers
-  const toggleMic = () => {
-    if (streamRef.current) {
-      const audioTrack = streamRef.current.getAudioTracks()[0];
-      if (audioTrack) {
-        audioTrack.enabled = !micOn;
-        setMicOn(!micOn);
-      }
-    }
-  };
-
-  const toggleCamera = () => {
-    if (streamRef.current) {
-      const videoTrack = streamRef.current.getVideoTracks()[0];
-      if (videoTrack) {
-        videoTrack.enabled = !cameraOn;
-        setCameraOn(!cameraOn);
-      }
-    }
-  };
-
-  const toggleScreenShare = async () => {
-    if (!isScreenSharing) {
-      try {
-        const screenStream = await navigator.mediaDevices.getDisplayMedia({
-          video: { cursor: true },
-        });
-        const screenTrack = screenStream.getVideoTracks()[0];
-
-        peersRef.current.forEach(({ peer }) => {
-          const videoTrack = streamRef.current.getVideoTracks()[0];
-          peer.replaceTrack(videoTrack, screenTrack, streamRef.current);
-        });
-
-        screenTrack.onended = () => {
-          stopScreenShare();
-        };
-
-        userVideoRef.current.srcObject = screenStream;
-        screenTrackRef.current = screenStream;
-        setIsScreenSharing(true);
-      } catch (err) {
-        console.error("Screen share failed", err);
-      }
-    } else {
-      stopScreenShare();
-    }
-  };
-
-  const stopScreenShare = () => {
-    const videoTrack = streamRef.current.getVideoTracks()[0];
-    peersRef.current.forEach(({ peer }) => {
-      const currentTrack = screenTrackRef.current?.getTracks()[0];
-      if (currentTrack) {
-        peer.replaceTrack(currentTrack, videoTrack, streamRef.current);
-      }
-    });
-
-    screenTrackRef.current?.getTracks().forEach((t) => t.stop());
-    userVideoRef.current.srcObject = streamRef.current;
-    setIsScreenSharing(false);
-  };
-
   const copyLink = () => {
     navigator.clipboard.writeText(window.location.href);
     toast.success("Meeting link copied!");
-  };
-
-  const startTranscription = async () => {
-    try {
-      socketRef.current.emit("start-transcription", { roomId });
-      setupAudioProcessing();
-    } catch (error) {
-      console.error("Failed to start transcription:", error);
-      toast.error("Failed to start transcription");
-    }
-  };
-
-  const stopTranscription = async () => {
-    try {
-      socketRef.current.emit("stop-transcription", { roomId });
-      if (audioProcessor) {
-        audioProcessor.disconnect();
-        setAudioProcessor(null);
-      }
-      if (audioContext) {
-        await audioContext.close();
-        setAudioContext(null);
-      }
-    } catch (error) {
-      console.error("Failed to stop transcription:", error);
-      toast.error("Failed to stop transcription");
-    }
-  };
-
-  const setupAudioProcessing = async () => {
-    try {
-      const context = new (window.AudioContext || window.webkitAudioContext)();
-      setAudioContext(context);
-
-      const source = context.createMediaStreamSource(streamRef.current);
-      const processor = context.createScriptProcessor(4096, 1, 1);
-
-      processor.onaudioprocess = (e) => {
-        const inputData = e.inputBuffer.getChannelData(0);
-        const audioData = new Float32Array(inputData);
-        const pcmData = new Int16Array(audioData.length);
-
-        for (let i = 0; i < audioData.length; i++) {
-          pcmData[i] = Math.max(-1, Math.min(1, audioData[i])) * 0x7fff;
-        }
-
-        socketRef.current.emit("audio-data", {
-          roomId,
-          audioData: pcmData.buffer,
-        });
-      };
-
-      source.connect(processor);
-      processor.connect(context.destination);
-      setAudioProcessor(processor);
-    } catch (error) {
-      console.error("Error setting up audio processing:", error);
-    }
-  };
-
-  const toggleTranscription = () => {
-    if (transcriptionEnabled) {
-      stopTranscription();
-    } else {
-      startTranscription();
-    }
   };
 
   return (
@@ -480,82 +124,18 @@ const MeetingRoom = () => {
       {/* ---------- ACTIVE MEETING SCREEN ---------- */}
       {joined && !meetingEnded && (
         <div className="flex-1 flex flex-col min-h-0 bg-gray-900 relative">
-          {/* Header */}
-          <div className="h-16 bg-gray-900 border-b border-gray-800 flex items-center justify-between px-6 z-20 shrink-0">
-            <div className="flex items-center gap-4">
-              <h2 className="text-lg font-bold text-white truncate max-w-xs md:max-w-md">
-                Room: {roomId}
-              </h2>
-              <div className="flex items-center gap-2 text-gray-300 bg-gray-800 px-3 py-1 rounded-full text-sm font-mono">
-                <Clock size={14} />
-                <span>{formatTime(duration)}</span>
-              </div>
-              <div className="flex items-center gap-2 text-gray-300 bg-gray-800 px-3 py-1 rounded-full text-sm">
-                <Users size={16} />
-                <span>{peers.length + 1}</span>
-              </div>
-            </div>
-
-            <button
-              onClick={copyLink}
-              className="text-gray-300 hover:text-white flex items-center gap-1.5 text-sm font-semibold bg-gray-800 hover:bg-gray-700 px-4 py-2 rounded-xl transition-all cursor-pointer"
-            >
-              <Copy size={16} />
-              <span className="hidden sm:inline">Copy Link</span>
-            </button>
-
-            {/* Notes Toggle */}
-            <button
-              onClick={() => setShowNotes((v) => !v)}
-              className={`flex items-center gap-1.5 text-sm font-semibold px-4 py-2 rounded-xl transition-all cursor-pointer ${
-                showNotes
-                  ? "bg-indigo-600 text-white hover:bg-indigo-700"
-                  : "bg-gray-800 text-gray-300 hover:text-white hover:bg-gray-700"
-              }`}
-              title={showNotes ? "Hide notes" : "Open collaborative notes"}
-            >
-              {showNotes ? (
-                <PanelRightClose size={16} />
-              ) : (
-                <NotebookPen size={16} />
-              )}
-              <span className="hidden sm:inline">
-                {showNotes ? "Hide Notes" : "Notes"}
-              </span>
-            </button>
-
-            {/* Transcription Toggle */}
-            <button
-              onClick={toggleTranscription}
-              className={`flex items-center gap-1.5 text-sm font-semibold px-4 py-2 rounded-xl transition-all cursor-pointer ${
-                transcriptionEnabled
-                  ? "bg-green-600 text-white hover:bg-green-700"
-                  : "bg-gray-800 text-gray-300 hover:text-white hover:bg-gray-700"
-              }`}
-              title={transcriptionEnabled ? "Stop transcription" : "Start live transcription"}
-            >
-              <Captions size={16} />
-              <span className="hidden sm:inline">
-                {transcriptionEnabled ? "Stop" : "Captions"}
-              </span>
-            </button>
-
-            {/* Transcript Toggle */}
-            <button
-              onClick={() => setShowTranscript((v) => !v)}
-              className={`flex items-center gap-1.5 text-sm font-semibold px-4 py-2 rounded-xl transition-all cursor-pointer ${
-                showTranscript
-                  ? "bg-indigo-600 text-white hover:bg-indigo-700"
-                  : "bg-gray-800 text-gray-300 hover:text-white hover:bg-gray-700"
-              }`}
-              title={showTranscript ? "Hide transcript" : "Show transcript"}
-            >
-              <FileText size={16} />
-              <span className="hidden sm:inline">
-                {showTranscript ? "Hide" : "Transcript"}
-              </span>
-            </button>
-          </div>
+          <MeetingHeader
+            roomId={roomId}
+            duration={duration}
+            peers={peers}
+            copyLink={copyLink}
+            showNotes={showNotes}
+            setShowNotes={setShowNotes}
+            transcriptionEnabled={transcriptionEnabled}
+            toggleTranscription={toggleTranscription}
+            showTranscript={showTranscript}
+            setShowTranscript={setShowTranscript}
+          />
 
           {/* Main content area: video grid + notes panel */}
           <div className="flex-1 flex min-h-0 overflow-hidden">
@@ -609,130 +189,24 @@ const MeetingRoom = () => {
             )}
 
             {/* Transcript Panel */}
-            {showTranscript && (
-              <div className="w-full md:w-[420px] lg:w-[480px] shrink-0 p-4 bg-gray-950 border-l border-gray-800 overflow-hidden flex flex-col">
-                <div className="flex items-center justify-between mb-4">
-                  <h3 className="text-white font-semibold flex items-center gap-2">
-                    <FileText size={18} />
-                    Live Transcript
-                  </h3>
-                  <button
-                    onClick={() => setShowTranscript(false)}
-                    className="text-gray-400 hover:text-white transition-colors"
-                  >
-                    <X size={20} />
-                  </button>
-                </div>
-                <div className="flex-1 overflow-y-auto space-y-3">
-                  {transcriptSegments.length === 0 ? (
-                    <div className="text-gray-500 text-center py-8">
-                      <Captions size={32} className="mx-auto mb-2 opacity-50" />
-                      <p className="text-sm">No transcript yet</p>
-                      <p className="text-xs mt-1">Enable captions to start transcription</p>
-                    </div>
-                  ) : (
-                    transcriptSegments.map((segment, index) => (
-                      <div
-                        key={index}
-                        className="bg-gray-800 rounded-lg p-3 border border-gray-700"
-                      >
-                        <div className="flex items-center justify-between mb-1">
-                          <span className="text-indigo-400 text-sm font-medium">
-                            {segment.speaker || "Speaker"}
-                          </span>
-                          <span className="text-gray-500 text-xs">
-                            {formatTimestamp(segment.startTime)}
-                          </span>
-                        </div>
-                        <p className="text-gray-300 text-sm">{segment.text}</p>
-                      </div>
-                    ))
-                  )}
-                </div>
-              </div>
-            )}
+            <TranscriptPanel
+              showTranscript={showTranscript}
+              setShowTranscript={setShowTranscript}
+              transcriptSegments={transcriptSegments}
+            />
           </div>
 
-          {/* Live Caption Bar */}
-          {showCaptions && captions.length > 0 && (
-            <div className="bg-gray-800/90 backdrop-blur-sm border-t border-gray-700 px-6 py-3">
-              <div className="flex items-center gap-2 mb-2">
-                <Captions size={16} className="text-indigo-400" />
-                <span className="text-gray-400 text-xs font-medium">Live Captions</span>
-              </div>
-              <div className="space-y-1 max-h-24 overflow-y-auto">
-                {captions.map((caption, index) => (
-                  <div
-                    key={index}
-                    className={`text-sm ${
-                      caption.isFinal ? "text-white" : "text-gray-400 italic"
-                    }`}
-                  >
-                    {caption.speaker && (
-                      <span className="text-indigo-400 font-medium mr-2">
-                    {caption.speaker}:
-                  </span>
-                )}
-                    {caption.text}
-                  </div>
-                ))}
-              </div>
-            </div>
-          )}
+          <LiveCaptions showCaptions={showCaptions} captions={captions} />
 
-          {/* Control Bar */}
-          <div className="h-24 bg-gray-900 border-t border-gray-800 flex items-center justify-center gap-4 px-6 z-20 shrink-0">
-            {/* Mic Toggle */}
-            <button
-              onClick={toggleMic}
-              className={`p-4 rounded-full transition-all shadow-md active:scale-95 cursor-pointer ${
-                micOn
-                  ? "bg-gray-800 text-white hover:bg-gray-700"
-                  : "bg-red-500 text-white hover:bg-red-600"
-              }`}
-              aria-label={micOn ? "Mute microphone" : "Unmute microphone"}
-            >
-              {micOn ? <Mic size={22} /> : <MicOff size={22} />}
-            </button>
-
-            {/* Camera Toggle */}
-            <button
-              onClick={toggleCamera}
-              className={`p-4 rounded-full transition-all shadow-md active:scale-95 cursor-pointer ${
-                cameraOn
-                  ? "bg-gray-800 text-white hover:bg-gray-700"
-                  : "bg-red-500 text-white hover:bg-red-600"
-              }`}
-              aria-label={cameraOn ? "Turn off camera" : "Turn on camera"}
-            >
-              {cameraOn ? <Video size={22} /> : <VideoOff size={22} />}
-            </button>
-
-            {/* Screen Share */}
-            <button
-              onClick={toggleScreenShare}
-              className={`p-4 rounded-full transition-all shadow-md active:scale-95 cursor-pointer ${
-                isScreenSharing
-                  ? "bg-indigo-500 text-white hover:bg-indigo-600"
-                  : "bg-gray-800 text-white hover:bg-gray-700"
-              }`}
-              aria-label={
-                isScreenSharing ? "Stop screen share" : "Share screen"
-              }
-            >
-              <MonitorUp size={22} />
-            </button>
-
-            <div className="w-px h-8 bg-gray-700 mx-2"></div>
-
-            {/* Leave */}
-            <button
-              onClick={leaveMeeting}
-              className="px-6 py-4 bg-red-600 text-white rounded-full font-semibold hover:bg-red-700 shadow-lg transition-all active:scale-95 flex items-center gap-2 cursor-pointer"
-            >
-              <PhoneOff size={22} /> Leave
-            </button>
-          </div>
+          <MeetingControlBar
+            micOn={micOn}
+            toggleMic={toggleMic}
+            cameraOn={cameraOn}
+            toggleCamera={toggleCamera}
+            isScreenSharing={isScreenSharing}
+            toggleScreenShare={toggleScreenShare}
+            leaveMeeting={leaveMeeting}
+          />
         </div>
       )}
