+++ client/src/utils/taskStyles.js
@@ -1,0 +1,56 @@
+import { CheckCircle2, Clock, AlertCircle, Loader2 } from "lucide-react";
+
+export const STATUS_STYLES = {
+  open: {
+    label: "Open",
+    bgColor: "bg-slate-100 dark:bg-slate-800",
+    textColor: "text-slate-700 dark:text-slate-300",
+    borderColor: "border-slate-200 dark:border-slate-700",
+    icon: Clock,
+  },
+
+  "in-progress": {
+    label: "In Progress",
+    bgColor: "bg-blue-50 dark:bg-blue-900/30",
+    textColor: "text-blue-700 dark:text-blue-400",
+    borderColor: "border-blue-200 dark:border-blue-800",
+    icon: Loader2,
+  },
+
+  resolved: {
+    label: "Resolved",
+    bgColor: "bg-emerald-50 dark:bg-emerald-900/30",
+    textColor: "text-emerald-700 dark:text-emerald-400",
+    borderColor: "border-emerald-200 dark:border-emerald-800",
+    icon: CheckCircle2,
+  },
+
+  superseded: {
+    label: "Superseded",
+    bgColor: "bg-red-50 dark:bg-red-900/30",
+    textColor: "text-red-700 dark:text-red-400",
+    borderColor: "border-red-200 dark:border-red-800",
+    icon: AlertCircle,
+  },
+};
+
+export const PRIORITY_STYLES = {
+  high: {
+    label: "High",
+    bgColor: "bg-red-100 dark:bg-red-900/30",
+    textColor: "text-red-700 dark:text-red-400",
+    borderColor: "border-red-200 dark:border-red-800",
+  },
+  medium: {
+    label: "Medium",
+    bgColor: "bg-amber-100 dark:bg-amber-900/30",
+    textColor: "text-amber-700 dark:text-amber-400",
+    borderColor: "border-amber-200 dark:border-amber-800",
+  },
+  low: {
+    label: "Low",
+    bgColor: "bg-slate-100 dark:bg-slate-800",
+    textColor: "text-slate-700 dark:text-slate-300",
+    borderColor: "border-slate-200 dark:border-slate-700",
+  },
+};
